Final cleanup for new funtional test fixtures
