+++ securedrop/tests/functional/app_navigators/_nav_helper.py
@@ -1,0 +1,140 @@
+import time
+from typing import List
+
+from selenium.common.exceptions import NoAlertPresentException, WebDriverException
+from selenium.webdriver.common.by import By
+from selenium.webdriver.firefox.webdriver import WebDriver
+from selenium.webdriver.remote.webelement import WebElement
+from selenium.webdriver.support import expected_conditions
+from selenium.webdriver.support.wait import WebDriverWait
+
+
+class NavigationHelper:
+
+    _TIMEOUT = 10
+    _POLL_FREQUENCY = 0.1
+
+    def __init__(self, web_driver: WebDriver) -> None:
+        self.driver = web_driver
+
+    def wait_for(self, function_with_assertion, timeout=_TIMEOUT):
+        """Polling wait for an arbitrary assertion."""
+        # Thanks to
+        # http://chimera.labs.oreilly.com/books/1234000000754/ch20.html#_a_common_selenium_problem_race_conditions
+        start_time = time.time()
+        while time.time() - start_time < timeout:
+            try:
+                return function_with_assertion()
+            except (AssertionError, WebDriverException):
+                time.sleep(self._POLL_FREQUENCY)
+        # one more try, which will raise any errors if they are outstanding
+        return function_with_assertion()
+
+    def safe_click_by_id(self, element_id: str) -> WebElement:
+        """
+        Clicks the element with the given ID attribute.
+
+        Returns:
+            el: The element, if found.
+
+        Raises:
+            selenium.common.exceptions.TimeoutException: If the element cannot be found in time.
+
+        """
+        el = WebDriverWait(self.driver, self._TIMEOUT, self._POLL_FREQUENCY).until(
+            expected_conditions.element_to_be_clickable((By.ID, element_id))
+        )
+        el.location_once_scrolled_into_view
+        el.click()
+        return el
+
+    def safe_click_by_css_selector(self, selector: str) -> WebElement:
+        """
+        Clicks the first element with the given CSS selector.
+
+        Returns:
+            el: The element, if found.
+
+        Raises:
+            selenium.common.exceptions.TimeoutException: If the element cannot be found in time.
+
+        """
+        el = WebDriverWait(self.driver, self._TIMEOUT, self._POLL_FREQUENCY).until(
+            expected_conditions.element_to_be_clickable((By.CSS_SELECTOR, selector))
+        )
+        el.click()
+        return el
+
+    def safe_click_all_by_css_selector(self, selector: str) -> List[WebElement]:
+        """
+        Clicks each element that matches the given CSS selector.
+
+        Returns:
+            els (list): The list of elements that matched the selector.
+
+        Raises:
+            selenium.common.exceptions.TimeoutException: If the element cannot be found in time.
+
+        """
+        els = self.wait_for(lambda: self.driver.find_elements_by_css_selector(selector))
+        for el in els:
+            clickable_el = WebDriverWait(self.driver, self._TIMEOUT, self._POLL_FREQUENCY).until(
+                expected_conditions.element_to_be_clickable((By.CSS_SELECTOR, selector))
+            )
+            clickable_el.click()
+        return els
+
+    def safe_send_keys_by_id(self, element_id: str, text: str) -> WebElement:
+        """
+        Sends the given text to the element with the specified ID.
+
+        Returns:
+            el: The element, if found.
+
+        Raises:
+            selenium.common.exceptions.TimeoutException: If the element cannot be found in time.
+
+        """
+        el = WebDriverWait(self.driver, self._TIMEOUT, self._POLL_FREQUENCY).until(
+            expected_conditions.element_to_be_clickable((By.ID, element_id))
+        )
+        el.send_keys(text)
+        return el
+
+    def safe_send_keys_by_css_selector(self, selector: str, text: str) -> WebElement:
+        """
+        Sends the given text to the first element with the given CSS selector.
+
+        Returns:
+            el: The element, if found.
+
+        Raises:
+            selenium.common.exceptions.TimeoutException: If the element cannot be found in time.
+
+        """
+        el = WebDriverWait(self.driver, self._TIMEOUT, self._POLL_FREQUENCY).until(
+            expected_conditions.element_to_be_clickable((By.CSS_SELECTOR, selector))
+        )
+        el.send_keys(text)
+        return el
+
+    def alert_wait(self, timeout: int = _TIMEOUT * 10) -> None:
+        WebDriverWait(self.driver, timeout, self._POLL_FREQUENCY).until(
+            expected_conditions.alert_is_present(), "Timed out waiting for confirmation popup."
+        )
+
+    def alert_accept(self) -> None:
+        # adapted from https://stackoverflow.com/a/34795883/837471
+        def alert_is_not_present(object):
+            """Expect an alert to not be present."""
+            try:
+                alert = self.driver.switch_to.alert
+                alert.text
+                return False
+            except NoAlertPresentException:
+                return True
+
+        self.driver.switch_to.alert.accept()
+        WebDriverWait(self.driver, self._TIMEOUT, self._POLL_FREQUENCY).until(
+            alert_is_not_present, "Timed out waiting for confirmation popup to disappear."
+        )
