+++ securedrop/tests/functional/pageslayout/test_source_static_pages.py
@@ -1,8 +1,7 @@
 import pytest
 import requests
 from tests.functional import tor_utils
-from tests.functional.pageslayout.functional_test import list_locales
-from tests.functional.pageslayout.screenshot_utils import save_screenshot_and_html
+from tests.functional.pageslayout.utils import list_locales, save_screenshot_and_html
 from tests.functional.web_drivers import WebDriverTypeEnum, get_web_driver
 from version import __version__
 
@@ -10,7 +9,7 @@ from version import __version__
 @pytest.mark.pagelayout
 class TestSourceAppStaticPages:
     @pytest.mark.parametrize("locale", list_locales())
-    def test_notfound(self, locale, sd_servers_v2):
+    def test_notfound(self, locale, sd_servers):
         # Given a source user accessing the app from their browser
         locale_with_commas = locale.replace("_", "-")
         with get_web_driver(
@@ -19,7 +18,7 @@ class TestSourceAppStaticPages:
         ) as tor_browser_web_driver:
 
             # When they try to access a page that does not exist
-            tor_browser_web_driver.get(f"{sd_servers_v2.source_app_base_url}/does_not_exist")
+            tor_browser_web_driver.get(f"{sd_servers.source_app_base_url}/does_not_exist")
 
             # Then the right error is displayed
             message = tor_browser_web_driver.find_element_by_id("page-not-found")
@@ -28,7 +27,7 @@ class TestSourceAppStaticPages:
             save_screenshot_and_html(tor_browser_web_driver, locale, "source-notfound")
 
     @pytest.mark.parametrize("locale", list_locales())
-    def test_static_pages(self, locale, sd_servers_v2):
+    def test_static_pages(self, locale, sd_servers):
         # Given a source user accessing the app from their browser
         locale_with_commas = locale.replace("_", "-")
         with get_web_driver(
@@ -36,18 +35,18 @@ class TestSourceAppStaticPages:
             accept_languages=locale_with_commas,
         ) as tor_browser_web_driver:
             # The user can browse to some of the app's static pages
-            tor_browser_web_driver.get(f"{sd_servers_v2.source_app_base_url}/use-tor")
+            tor_browser_web_driver.get(f"{sd_servers.source_app_base_url}/use-tor")
             save_screenshot_and_html(tor_browser_web_driver, locale, "source-use_tor_browser")
 
-            tor_browser_web_driver.get(f"{sd_servers_v2.source_app_base_url}/tor2web-warning")
+            tor_browser_web_driver.get(f"{sd_servers.source_app_base_url}/tor2web-warning")
             save_screenshot_and_html(tor_browser_web_driver, locale, "source-tor2web_warning")
 
-            tor_browser_web_driver.get(f"{sd_servers_v2.source_app_base_url}/why-public-key")
+            tor_browser_web_driver.get(f"{sd_servers.source_app_base_url}/why-public-key")
             save_screenshot_and_html(tor_browser_web_driver, locale, "source-why_journalist_key")
 
-    def test_instance_metadata(self, sd_servers_v2):
+    def test_instance_metadata(self, sd_servers):
         # Given a source app, when fetching the instance's metadata
-        url = f"{sd_servers_v2.source_app_base_url}/metadata"
+        url = f"{sd_servers.source_app_base_url}/metadata"
         response = requests.get(url=url, proxies=tor_utils.proxies_for_url(url))
 
         # Then it succeeds and the right information is returned
